backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
aiogram
httpx